% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{chartProgress}
\alias{chartProgress}
\title{Chart the Student's Progress}
\usage{
chartProgress(students)
}
\arguments{
\item{students}{A data.table, the output of \link{getStudentsProgress}}
}
\description{
Chart the Student's Progress
}
