% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose-headers.R
\name{composeNexusHandle}
\alias{composeNexusHandle}
\title{Compose a Nexus Curl Handle}
\usage{
composeNexusHandle(netid)
}
\arguments{
\item{netid}{The mentor netid}
}
\value{
A handle object (external pointer to the underlying curl handle).
}
\description{
Compose a Nexus Curl Handle
}
\examples{
\dontrun{
h <- composeNexusHandle("mkemp6")
}
}
