% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose-headers.R
\name{decrpytChromeCookies}
\alias{decrpytChromeCookies}
\title{Decrypt the Chrome Cookies for Nexus}
\usage{
decrpytChromeCookies(netid)
}
\arguments{
\item{netid}{The mentor netid}
}
\value{
a \code{\link[data.table]{data.table}} with the cookie names and values
    along with other meta information.
}
\description{
Decrypt the Chrome Cookies for Nexus
}
