% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-extractors.R
\name{extractAssignment}
\alias{extractAssignment}
\title{Extract Important Assignment Information from JSON List}
\usage{
extractAssignment(assignment)
}
\arguments{
\item{assignment}{A JSON list}
}
\value{
A \code{\link[data.table]{data.table}}
}
\description{
Extract Important Assignment Information from JSON List
}
