% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-extractors.R
\name{extractLatestEmailDate}
\alias{extractLatestEmailDate}
\title{Extract the Latest Email From Student}
\usage{
extractLatestEmailDate(nexus_cookies = NULL, student_netid)
}
\arguments{
\item{nexus_cookies}{A data.table of cookies for Nexus' \code{sessionId} (and any others)}

\item{student_netid}{A student's NetId}
}
\value{
A Date type object
}
\description{
Extract the Latest Email From Student
}
