% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose-headers.R
\name{extractMathableCookies}
\alias{extractMathableCookies}
\title{Compose a Mathable Curl Handle}
\usage{
extractMathableCookies(netid)
}
\arguments{
\item{netid}{The user netid}
}
\value{
A handle object (external pointer to the underlying curl handle).
}
\description{
Compose a Mathable Curl Handle
}
