% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose-headers.R
\name{extractNexusCookies}
\alias{extractNexusCookies}
\title{Extract Nexus Cookie(s) from FireFox}
\usage{
extractNexusCookies(profile_pattern = "\\\\.default-release$")
}
\arguments{
\item{profile_pattern}{A regex pattern to help find the correct firefox profile from which to use the \code{cookies.sqlite} DB}
}
\value{
A data.table
}
\description{
Extract Nexus Cookie(s) from FireFox
}
