% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-extractors.R
\name{extractStudent}
\alias{extractStudent}
\title{Extract Important Student Information from JSON List}
\usage{
extractStudent(student, handle)
}
\arguments{
\item{student}{A JSON list}

\item{handle}{A \code{curl} \code{\link[curl]{handle}}}
}
\value{
A \code{\link[data.table]{data.table}}
}
\description{
Extract Important Student Information from JSON List
}
