% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-extractors.R
\name{extractStudentProgress}
\alias{extractStudentProgress}
\title{Extract Student's Progress in a Course}
\usage{
extractStudentProgress(notebooks, course_id, days_left)
}
\arguments{
\item{notebooks}{A JSON list}

\item{course_id}{A character string}

\item{days_left}{A numeric value}
}
\value{
A \code{\link[data.table]{data.table}} summarizing the student's progress
}
\description{
Extract Student's Progress in a Course
}
