% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentor-activity.R
\name{getAsssignmentDetail}
\alias{getAsssignmentDetail}
\title{Get the details about a Mathable assignment}
\usage{
getAsssignmentDetail(netid, course_id)
}
\arguments{
\item{course_id}{}
}
\description{
Get the details about a Mathable assignment
}
