% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{getGrades}
\alias{getGrades}
\title{Get a Student's Grades}
\usage{
getGrades(handle, student_id)
}
\arguments{
\item{handle}{An active session established with \link{composeNexusHandle}}

\item{student_id}{A character string}
}
\value{
A JSON list
}
\description{
Get a Student's Grades
}
