% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{getMathable}
\alias{getMathable}
\title{Execute a GET request}
\usage{
getMathable(netid, mathable_cookies = NULL, route, where = "nexus", ...)
}
\arguments{
\item{route}{A character string}

\item{...}{Additional arguments to be passed in the url}

\item{handle}{An active session established with \link{composeNexusHandle}}
}
\value{
A JSON list
}
\description{
Execute a GET request
}
