% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{getMathableCourse}
\alias{getMathableCourse}
\title{Get Mathable Course Information}
\usage{
getMathableCourse(netid, student_courseid)
}
\arguments{
\item{netid}{Mentor NetID}

\item{student_courseid}{Student's Mathable Course ID}
}
\description{
Get Mathable Course Information
}
