% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentor-activity.R
\name{getMentorActivity}
\alias{getMentorActivity}
\title{Get mentor email activity}
\usage{
getMentorActivity(
  students,
  nexus_cookies = NULL,
  earliest_date = lubridate::floor_date(Sys.Date(), "week") - 2 * 7
)
}
\arguments{
\item{nexus_cookies}{}
}
\value{

}
\description{
Get mentor email activity
}
