% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{getNexusRequest}
\alias{getNexusRequest}
\title{Execute a request to Nexus}
\usage{
getNexusRequest(route, nexus_cookies = NULL, ...)
}
\arguments{
\item{route}{What route? E.g. \code{students}, as a character string}

\item{nexus_cookies}{A data.table of cookies for Nexus' \code{sessionId} (and any others)}

\item{...}{Additional arguments to be passed in the url}
}
\value{

}
\description{
Execute a request to Nexus
}
