% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{getRequest}
\alias{getRequest}
\title{Execute a GET request}
\usage{
getRequest(handle, route, where = "nexus", ...)
}
\arguments{
\item{handle}{An active session established with \link{composeNexusHandle}}

\item{route}{A character string}

\item{...}{Additional arguments to be passed in the url}
}
\value{
A JSON list
}
\description{
Execute a GET request
}
