% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{getStudentsProgress}
\alias{getStudentsProgress}
\title{Get Students' Progress}
\usage{
getStudentsProgress(students, nexus_cookies = NULL)
}
\arguments{
\item{students}{A \code{\link[data.table]{data.table}} returned by \link{getStudents}}

\item{nexus_cookies}{A data.table of cookies for Nexus' \code{sessionId} (and any others)}
}
\value{
A \code{\link[data.table]{data.table}} with the same columns and some
    progress metrics appended
}
\description{
Get Students' Progress
}
