% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{getTicketList}
\alias{getTicketList}
\title{Request Student's Ticket Threads}
\usage{
getTicketList(student_netid, nexus_cookies = NULL, page = 1)
}
\arguments{
\item{student_netid}{Student's netId}

\item{nexus_cookies}{A data.table of cookies for Nexus' \code{sessionId} (and any others)}

\item{page}{Default \code{1}, the page of results to request}
}
\value{
A JSON list
}
\description{
Request Student's Ticket Threads
}
