% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{getTicketMessages}
\alias{getTicketMessages}
\title{Request a Specific Student Ticket}
\usage{
getTicketMessages(ticket_id, nexus_cookies = NULL)
}
\arguments{
\item{ticket_id}{The unique ticket_id of the}

\item{nexus_cookies}{A data.table of cookies for Nexus' \code{sessionId} (and any others)}
}
\value{
A JSON list
}
\description{
Request a Specific Student Ticket
}
