% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{putNexus}
\alias{putNexus}
\title{Execute a put to Nexus}
\usage{
putNexus(route, nexus_cookies = NULL, ...)
}
\arguments{
\item{route}{What route? E.g. \code{students}, as a character string}

\item{nexus_cookies}{A data.table of cookies for Nexus' \code{sessionId} (and any others)}

\item{...}{Additional arguments to be passed in the url}
}
\value{

}
\description{
Execute a put to Nexus
}
