% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose-headers.R
\name{composeNexusHandle}
\alias{composeNexusHandle}
\title{Compose the Nexus  the Chrome Cookies for Nexus}
\usage{
composeNexusHandle(netid)
}
\arguments{
\item{netid}{The mentor netid}
}
\value{
A handle object (external pointer to the underlying curl handle).
}
\description{
Compose the Nexus  the Chrome Cookies for Nexus
}
\examples{
\dontrun{
h <- composeNexusHandle("mkemp6")
}
}
