% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-extractors.R
\name{extractLatestEmailDate}
\alias{extractLatestEmailDate}
\title{Extract the Latest Email From Student}
\usage{
extractLatestEmailDate(handle, student_netid)
}
\arguments{
\item{handle}{An active session established with \link{composeNexusHandle}}

\item{student_netid}{A student's NetId}
}
\value{
A Date type object
}
\description{
Extract the Latest Email From Student
}
