% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-extractors.R
\name{extractStudent}
\alias{extractStudent}
\title{Extract Important Student Information from JSON List}
\usage{
extractStudent(student, nexus_cookies = NULL)
}
\arguments{
\item{student}{A JSON list}

\item{nexus_cookies}{A data.table of cookies for Nexus' \code{sessionId} (and any others)}
}
\value{
A \code{\link[data.table]{data.table}}
}
\description{
Extract Important Student Information from JSON List
}
