% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{getGrades}
\alias{getGrades}
\title{Get a Student's Grades}
\usage{
getGrades(student_id, nexus_cookies = NULL)
}
\arguments{
\item{student_id}{A character string}

\item{nexus_cookies}{A data.table of cookies for Nexus' \code{sessionId} (and any others)}
}
\value{
A JSON list
}
\description{
Get a Student's Grades
}
