% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{getMathableNotebook}
\alias{getMathableNotebook}
\title{Get the specifics of a notebook}
\usage{
getMathableNotebook(handle, student_netid, notebook_id)
}
\arguments{
\item{handle}{Valid Mathable curl handle}

\item{student_netid}{Student's Id}

\item{notebook_id}{Mathable Notebook Id}
}
\value{

}
\description{
Get the specifics of a notebook
}
