% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{getStudents}
\alias{getStudents}
\title{Get a Mentor's Students}
\usage{
getStudents(net_id, nexus_cookies = NULL)
}
\arguments{
\item{net_id}{A character string}

\item{nexus_cookies}{A data.table of cookies for Nexus' \code{sessionId} (and any others)}
}
\value{
A \code{\link[data.table]{data.table}}
}
\description{
Get a Mentor's Students
}
