% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{getStudents}
\alias{getStudents}
\title{Get a Mentor's Students}
\usage{
getStudents(handle, net_id)
}
\arguments{
\item{handle}{An active session established with \link{composeNexusHandle}}

\item{net_id}{A character string}
}
\value{
A \code{\link[data.table]{data.table}}
}
\description{
Get a Mentor's Students
}
