% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{getStudentsProgress}
\alias{getStudentsProgress}
\title{Get Students' Progress}
\usage{
getStudentsProgress(handle, students)
}
\arguments{
\item{handle}{An active session established with \link{composeNexusHandle}}

\item{students}{A \code{\link[data.table]{data.table}} returned by \link{getStudents}}
}
\value{
A \code{\link[data.table]{data.table}} with the same columns and some
    progress metrics appended
}
\description{
Get Students' Progress
}
