% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{getTicketList}
\alias{getTicketList}
\title{Request Student's Ticket Threads}
\usage{
getTicketList(handle, student_netid, page = 1)
}
\arguments{
\item{handle}{An active session established with \link{composeNexusHandle}}

\item{page}{Default \code{1}, the page of results to request}

\item{netId}{Student's netId}
}
\value{
A JSON list
}
\description{
Request Student's Ticket Threads
}
