% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{getTicketMessages}
\alias{getTicketMessages}
\title{Request a Specific Student Ticket}
\usage{
getTicketMessages(handle, ticket_id)
}
\arguments{
\item{handle}{An active session established with \link{composeNexusHandle}}

\item{ticket_id}{The unique ticket_id of the}
}
\value{
A JSON list
}
\description{
Request a Specific Student Ticket
}
